Move the deeply nested option logic to a toplevel folder.
