+++ packages/babel-core/src/config/option-manager.js
@@ -1,15 +1,15 @@
-import * as context from "../../../index";
+import * as context from "../index";
 import Plugin from "./plugin";
 import * as messages from "babel-messages";
-import resolvePlugin from "../../../helpers/resolve-plugin";
-import resolvePreset from "../../../helpers/resolve-preset";
+import resolvePlugin from "../helpers/resolve-plugin";
+import resolvePreset from "../helpers/resolve-preset";
 import defaults from "lodash/defaults";
 import cloneDeepWith from "lodash/cloneDeepWith";
-import merge from "../../../helpers/merge";
+import merge from "../helpers/merge";
 import removed from "./removed";
 import buildConfigChain from "./build-config-chain";
 import path from "path";
-import * as util from "../../../util";
+import * as util from "../util";
 
 type PluginObject = {
   pre?: Function;
