+++ packages/babel-core/src/config/plugin.js
