+++ packages/babel-core/src/config/removed.js
