+++ packages/babel-core/test/api.js
@@ -2,7 +2,7 @@ import * as babel from "../lib/index";
 import buildExternalHelpers from "../lib/tools/build-external-helpers";
 import sourceMap from "source-map";
 import assert from "assert";
-import Plugin from "../lib/transformation/file/options/plugin";
+import Plugin from "../lib/config/plugin";
 import generator from "babel-generator";
 
 function assertIgnored(result) {
